Allow multiple set of arguments to the partials
With this change the introduction of variable set of arguments is enabled so we can send multiple set of arguments. Every new set is merged with the initial set so at the end we end up with a single set with all the previous sets merged as with others frameworks such as `React` or `Object.assign` or `wp_parse_args`.